+++ packages/image_picker/image_picker_android/lib/image_picker_android.dart
@@ -135,6 +135,21 @@ class ImagePickerAndroid extends ImagePickerPlatform {
     return paths.isEmpty ? null : paths.first;
   }
 
+  Future<List<String>> _getMultiVideoPath({
+    Duration? maxDuration,
+    int? limit,
+  }) {
+    return _hostApi.pickVideos(
+      SourceSpecification(type: SourceType.gallery),
+      VideoSelectionOptions(maxDurationSeconds: maxDuration?.inSeconds),
+      GeneralOptions(
+        allowMultiple: true,
+        usePhotoPicker: useAndroidPhotoPicker,
+        limit: limit,
+      ),
+    );
+  }
+
   @override
   Future<PickedFile?> pickVideo({
     required ImageSource source,
@@ -259,6 +274,22 @@ class ImagePickerAndroid extends ImagePickerPlatform {
       preferredCameraDevice: preferredCameraDevice,
     );
     return path != null ? XFile(path) : null;
+  }
+
+  @override
+  Future<List<XFile>> getMultiVideoWithOptions({
+    MultiVideoPickerOptions options = const MultiVideoPickerOptions(),
+  }) async {
+    final List<String> paths = await _getMultiVideoPath(
+      maxDuration: options.maxDuration,
+      limit: options.limit,
+    );
+
+    if (paths.isEmpty) {
+      return <XFile>[];
+    }
+
+    return paths.map((String path) => XFile(path)).toList();
   }
 
   MediaSelectionOptions _mediaOptionsToMediaSelectionOptions(