+++ packages/image_picker/image_picker_platform_interface/lib/src/method_channel/method_channel_image_picker.dart
@@ -36,6 +36,21 @@ class MethodChannelImagePicker extends ImagePickerPlatform {
   }
 
   @override
+  Future<List<XFile>> getMultiVideoWithOptions({
+    MultiVideoPickerOptions options = const MultiVideoPickerOptions(),
+  }) async {
+    final List<dynamic>? paths = await _getMultiVideoPath(
+      maxDuration: options.maxDuration,
+      limit: options.limit,
+    );
+    if (paths == null) {
+      return <XFile>[];
+    }
+
+    return paths.map((dynamic path) => XFile(path as String)).toList();
+  }
+
+  @override
   Future<List<PickedFile>?> pickMultiImage({
     double? maxWidth,
     double? maxHeight,
@@ -80,6 +95,19 @@ class MethodChannelImagePicker extends ImagePickerPlatform {
         'maxHeight': maxHeight,
         'imageQuality': imageQuality,
         'requestFullMetadata': requestFullMetadata,
+        'limit': limit,
+      },
+    );
+  }
+
+  Future<List<dynamic>?> _getMultiVideoPath({
+    Duration? maxDuration,
+    int? limit,
+  }) {
+    return _channel.invokeMethod<List<dynamic>?>(
+      'pickMultiVideo',
+      <String, dynamic>{
+        'maxDuration': maxDuration?.inSeconds,
         'limit': limit,
       },
     );