+++ packages/image_picker/image_picker/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 1.2.0
 
+* Adds `pickMultiVideo` to allow selecting multiple videos from the gallery.
 * Updates minimum supported SDK version to Flutter 3.27/Dart 3.6.
 
 ## 1.1.2